+++ src/client/components/Cart.jsx
@@ -1,26 +1,38 @@
-import React from 'react';
+import React, { useState } from 'react';
+import NavBar from './NavBar';
 
 const handleAddToCart = (product) => {
-    // Create a copy of the current cartItems array and add the new product
-    const updatedCart = [...cartItems, product];
-  
-    // Update the cartItems state with the new array
-    setCartItems(updatedCart);
-  };
+  // Create a copy of the current cartItems array and add the new product
+  const updatedCart = [...cartItems, product];
 
+  // Update the cartItems state with the new array
+  cartItem(updatedCart);
+  
+};
   
 function Cart({ cartItems }) {
+
+  const cartItem = [];
+  function handleCart(e){
+    cartItems(e.target.value);
+    cartItem(cartItems);
+  }
+
+
   return (
-    <div>
+    <>
+    <NavBar />
+    <div className='outline padding'>
       <h2>Shopping Cart</h2>
       <ul>
-        {cartItems.map((item) => (
+        {cartItem.map((item) => (
           <li key={item.id}>
             {item.product_name} - {item.price}
           </li>
         ))}
       </ul>
     </div>
+    </>
   );
 }
