+++ src/client/components/NavBar.jsx
@@ -1,5 +1,6 @@
 import { Link } from "react-router-dom"
 
+
 export default function NavBar() {
 
 
@@ -7,13 +8,13 @@ export default function NavBar() {
         <>
         <div className="NavBar outline">
             <div>
-
-            </div>
-            <div>
                 <Link to = '/Login' className='navButton'>Login</Link>
                 <Link to = '/Register' className="navButton">Register</Link>
             </div>
             <Link to = '/' className="title">barely buy</Link>
+            <div>
+                <Link to = '/Cart' className="navButton">Cart</Link>
+            </div>
         </div>
         </>
     )
