+++ src/client/style.css
@@ -99,8 +99,8 @@ body{
     display:block;
     position:fixed;
     z-index: 99;
-    margin-left: 43%;
-    margin-top: 2%;
+    margin-left: 42.5vw;
+    margin-top: 2.5vh;
   }
 
   .padding {
